requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
pymongo==4.5.0
python-dotenv==1.0.0
Pillow==10.0.0
pandas==2.0.3
Werkzeug==2.3.7
email-validator==2.0.0.post2
pillow-heif==0.13.0
exifread==3.0.0
geopy==2.4.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
ultralytics==8.0.0
opencv-python-headless==4.8.0.74
matplotlib>=3.3
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
seaborn>=0.11.0 